refactor: 💡 (LocationSupplier) Location 영속화 후 반환하도록 변경
